Move tests to autofs_spec
Now that all the params are placed in the init.pp and the
package.pp and service.pp files look to the params in the init.pp,
the old tests were no longer passing. I removed the package_spec
and service_spec. Any tests in those files not already covered
in the autofs_spec have been added to the autofs_spec.