Update tests to use rspec-puppet-facts
The autofs_spec tests now use on_supported_os.each do from the
rspec-puppet-facts gem. This was required to prevent errors
related to not mocking osfamily or operatingsystem. It also
improves the scope for which the tests run as the validate against
all supported OS facts instead of just a manually entered subset