Update map and mount defined types to address #59
Moved all mapfile generation to the map defined type provided by
@traylenator and instead have the mount defined type calling and passing
values to the map defined type (if mapfile is not empty).

This allows all current functionality to continue working, as well as,
allowing the map defined type to be used standalone AND maintain the
existing mapfile generation feature set.

Tests and templates have been updated to account for the changes.