Ensure drop-in files non-executable
The drop-in directory, e.g. /etc/auto.master.d, is for files whose
content is intended to be interpreted as if it was part of the master
map.  It is not sensible for such files to be exectuable, but they were
being made executable when associated with an executable map file.
This is now fixed.