Fixed handling for bare-string mapcontents
The data type for the 'macontents' parameters of autofs::mount and
autofs::map permit both arrays and bare strings, and the documentation
is consistent with that, but catalog building failed when a bare string
is assigned to one of these parameters.  The bug was in the templates;
these are updated to handle the situation, and a unit test is added to
verify that it works.