Factored map file management out of autofs::mount
Performed a wide-reaching refactoring to relieve autofs::mount of the
responsibility of managing map files.  It now manages only entries in
the master map.  Created a new autofs::mapfile resource specific to
managing map files, as autofs::map has the wrong structure for this.
Created a new autofs::mapping resource to allow map file management to
be performed collaboratively via multiple resources, as autofs::map is
wrong for this, too.

To look at it a different way, this refactoring takes the space covered
by autofs::mount and autofs::map, and spreads it among a more targeted
autofs::mount, and new autofs::mapfile and autofs::mapping.  Autofs::map
is deprecated, but not removed.