Fixed issues revealed by CI testing
Fixed some issues that were revealed by failing CI tests and by the
process of fixing those.  Primarily, ensured that the Concat resource
for the master map is always declared if any autofs::mount is declared,
directly or indirectly.  This is the original behavior, which had
inadvertently been changed.  Also tweaked the format of some tests
that used the 'match' matcher, and changed the formats of a couple of
resource titles.