Removed deprecated params from autofs::mount
The several parameters of autofs::mount that had been made ineffective
and and flagged with deprecation warnings are now removed altogether.
Most of these were related to map file management, which autofs::mount
no longer does.  The removals are documented in README.md, and they
required a few adjustments to test code and data.  Took the opportunity
to clean some unused cruft out of the test data.