Add support for FreeBSD
FreeBSD 10.1 added support for an autofs implementation that is similar
to what this module manage with a few minor specificities from the Linux
world:

"autofs" is the actual driver of the automounter infrastructure in the
kernel.  It communicate with the "automountd" service when a device must
be mounted, features an "autounmountd" service to unmount unused
devices, and has a rc-script "automount" to update the kernel view of
the available mountpoints.