Quote and escape special chars in mapping keys
Previously, a mapping key could not include spaces because of the data
type.  Some implementations of autofs do support such keys when they are
properly quoted so do not limit their use with the module and properly
quote the key when it has special characters.

These chars still need to be manualy espaced when used with the fs
parameter because it is managed as a single String. This will be
addressed in a follow-up commit as it will be a breaking change.